Move ModifierExt into sample - Don't need these as Apis in Circuit proper
